// sportsWeather/src/app/models/player.ts
import { Team } from './team';
import { Game } from './game';

// export class Player {
//    firstName: String;
//    jerseyNum: String;
//    lastName: String;
//    teamId: Number;
//    games: Game[];
//    comments: Comment[];
// }

export class Player {
  id: Number;
  firstName: String;
  jerseyNum: String;
  lastName: String;
  type: String;
  likes: Number;
}